pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "staq-zeen"
version = "0.1.0"
description = "AI Financial Copilot with OPA Risk Gates and LangGraph Agent Runtime"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.6,<3.0",
    "fastapi>=0.115,<1.0",
    "httpx>=0.27,<1.0",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "langgraph>=0.2,<1.0",
    "langchain-core>=0.3,<1.0",
    "langchain-anthropic>=0.2,<1.0",
    "supabase>=2.5,<3.0",
    "e2b-code-interpreter>=1.0,<2.0",
    "mcp>=1.0,<2.0",
    "pgvector>=0.3,<1.0",
    "tiktoken>=0.7,<1.0",
    "opentelemetry-exporter-otlp>=1.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "respx>=0.21",
    "pytest-timeout>=2.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
